requirements.txt
anglerfish==2.5.0
css-html-js-minify==2.2.2
Jinja2==2.10
MarkupSafe==1.0
PyYAML==3.12
